requirements.txt
# Training dependencies (workstation with GPU)
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
fastai>=2.7.0
timm>=0.9.0
soundfile>=0.12.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
